api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
supabase==2.9.1
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.12
sse-starlette==2.1.3
anthropic==0.39.0
# Bridge core (chat, voice, action executor). Path is relative to repo root,
# which is where Render's build runs `pip install -r api/requirements.txt`.
-e ./bridge-core
